+++ test/test_formatters.py
@@ -124,6 +124,7 @@ Observation(
         Comment(id=2071611, username='borisb', created_at='Sep 05, 2018', truncated_body='suspect L. bardanae - but sits on Solanum (non-...')
     ],
     faves=[],
+    flags=[],
     identifications=[
         Identification(id=34896306, username='niconoe', taxon_name='Genus Lixus', created_at='Sep 05, 2018', truncated_body=''),
         Identification(id=34926789, username='borisb', taxon_name='Lixus bardanae', created_at='Sep 05, 2018', truncated_body=''),